move config tests to config_test.go
